+++ src/stateEnt/virtual/base/intervalEvent/intervalEvent.cpp
@@ -19,9 +19,9 @@
 
 #include "intervalEvent.h"
 
-IECBArg::IECBArg(unsigned long long e, int c, int m) : elapsedMs(e), cbCnt(c), maxCbCnt(m) {}
+IECBArg::IECBArg(unsigned long e, int c, int m) : elapsedMs(e), cbCnt(c), maxCbCnt(m) {}
 
-unsigned long long IECBArg::getElapsedMs()
+unsigned long IECBArg::getElapsedMs()
 {
   return elapsedMs;
 }
@@ -42,21 +42,21 @@ IntervalEvent::IntervalEvent()
   maxCbCnt = 0;
 }
 
-IntervalEvent::IntervalEvent(unsigned long long i, interval_event_cb c)
+IntervalEvent::IntervalEvent(unsigned long i, interval_event_cb c)
 {
   intervalMs = i;
   cb = c;
   maxCbCnt = MAX_CB_CNT_INF;
 }
 
-IntervalEvent::IntervalEvent(unsigned long long i, interval_event_cb c, int m)
+IntervalEvent::IntervalEvent(unsigned long i, interval_event_cb c, int m)
 {
   intervalMs = i;
   cb = c;
   maxCbCnt = m;
 }
 
-IntervalEvent::IntervalEvent(unsigned long long i, interval_event_cb c, int m, bool t)
+IntervalEvent::IntervalEvent(unsigned long i, interval_event_cb c, int m, bool t)
 {
   intervalMs = i;
   cb = c;
@@ -64,7 +64,7 @@ IntervalEvent::IntervalEvent(unsigned long long i, interval_event_cb c, int m, b
   transitory = t;
 }
 
-IntervalEvent::IntervalEvent(unsigned long long i, interval_event_cb c, int m, bool t, int m2)
+IntervalEvent::IntervalEvent(unsigned long i, interval_event_cb c, int m, bool t, int m2)
 {
   intervalMs = i;
   cb = c;
@@ -73,7 +73,7 @@ IntervalEvent::IntervalEvent(unsigned long long i, interval_event_cb c, int m, b
   mode = m2;
 }
 
-unsigned long long IntervalEvent::getIntervalMs()
+unsigned long IntervalEvent::getIntervalMs()
 {
   return intervalMs;
 }
@@ -93,12 +93,12 @@ int IntervalEvent::getCbCnt()
   return cbCnt;
 }
 
-unsigned long long IntervalEvent::getLastCbMs()
+unsigned long IntervalEvent::getLastCbMs()
 {
   return intervalMs * cbCnt;
 }
 
-unsigned long long IntervalEvent::getNextCbMs()
+unsigned long IntervalEvent::getNextCbMs()
 {
   return getLastCbMs() + intervalMs;
 }
@@ -108,12 +108,12 @@ bool IntervalEvent::getTransitory()
   return transitory;
 }
 
-bool IntervalEvent::isTime(unsigned long long elapsedMs)
+bool IntervalEvent::isTime(unsigned long elapsedMs)
 {
   return elapsedMs >= getNextCbMs() && (maxCbCnt < 0 || cbCnt < maxCbCnt);
 }
 
-bool IntervalEvent::cbIfTimeAndActive(unsigned long long elapsedMs)
+bool IntervalEvent::cbIfTimeAndActive(unsigned long elapsedMs)
 {
   if (mode == IE_MODE_ACTIVE && isTime(elapsedMs) && intervalMs && cb(IECBArg(elapsedMs, cbCnt, maxCbCnt))) // Checking intervalMs here since default constructor doesnt even define cb; might need stub there to be safe
   {
